+++ src/cds/graph/__init__.py
@@ -2,11 +2,16 @@
 
 from cds.graph.algorithms import (
     Graph,
+    bellman_ford,
     bfs,
+    connected_components,
+    degree,
     dfs,
     dijkstra,
+    floyd_warshall,
     has_cycle,
     kruskal_mst,
+    prim_mst,
     topological_sort,
 )
 
@@ -18,4 +23,9 @@ __all__ = [
     "kruskal_mst",
     "topological_sort",
     "has_cycle",
+    "degree",
+    "connected_components",
+    "bellman_ford",
+    "floyd_warshall",
+    "prim_mst",
 ]