+++ src/cds/montecarlo/__init__.py
@@ -3,6 +3,8 @@
 from cds.montecarlo.methods import (
     buffon_needle,
     estimate_pi,
+    hit_or_miss,
+    mc_expectation,
     mc_integrate,
     random_walk_1d,
     random_walk_2d,
@@ -11,6 +13,8 @@ from cds.montecarlo.methods import (
 __all__ = [
     "estimate_pi",
     "mc_integrate",
+    "mc_expectation",
+    "hit_or_miss",
     "random_walk_1d",
     "random_walk_2d",
     "buffon_needle",